+++ stress-tests/data-structures/RMQ.cpp
@@ -8,10 +8,10 @@ int main() {
 		vi v(N);
 		rep(i,0,N) v[i] = i;
 		random_shuffle(all(v));
-		RMQ<int> rmq(v);
+		RMQ<ll> rmq(v);
 		rep(i,0,N) rep(j,i+1,N+1) {
-			int m = rmq.query(i,j);
-			int n = 1 << 29;
+			ll m = rmq.query(i,j);
+			ll n = 1 << 29;
 			rep(k,i,j) n = min(n, v[k]);
 			assert(n == m);
 		}
