+++ stress-tests/graph/MaximalCliques.cpp
@@ -8,7 +8,7 @@ void fastCliques(vector<B>& eds, F f) {
 	vi deg(sz(eds));
 	rep(i,0,sz(eds)) deg[i] = sz(eds[i]);
 	rep(j,0,sz(eds)) {
-		int i = (int)(min_element(all(deg)) - deg.begin());
+		ll i = (ll)(min_element(all(deg)) - deg.begin());
 		R[i] = 1;
 		rec(eds, R, P & eds[i], X & eds[i], f);
 		R[i] = P[i] = 0; X[i] = 1;
@@ -21,13 +21,13 @@ int main1() {
 	rep(n,1,11) rep(m,0,200) {
 		vector<B> ed(n);
 		rep(i,0,m) {
-			int a = rand() % n, b = rand() % n;
+			ll a = rand() % n, b = rand() % n;
 			if (a == b) continue;
 			ed[a][b] = 1;
 			ed[b][a] = 1;
 		}
 		unordered_set<B> cl;
-		int co = 0;
+		ll co = 0;
 		cliques(ed, [&](B x) {
 			co++;
 			cl.insert(x);
@@ -49,7 +49,7 @@ int main1() {
 			assert(isClique(c));
 		}
 
-		int realCo = 0;
+		ll realCo = 0;
 		rep(bi,0,(1 << n)) {
 			B c{};
 			rep(i,0,n) c[i] = !!(bi & (1 << i));
@@ -63,16 +63,16 @@ int main1() {
 
 int main2() {
 	rep(it,0,20) {
-		const int n = 128, m = 4000;
+		const ll n = 128, m = 4000;
 		vector<B> ed(n);
 		rep(i,0,m) {
-			int a = rand() % n, b = rand() % n;
+			ll a = rand() % n, b = rand() % n;
 			if (a == b) continue;
 			ed[a][b] = 1;
 			ed[b][a] = 1;
 		}
-		int co = 0, sum = 0;
-		cliques(ed, [&](B x) { co++; sum += (int)x.count(); });
+		ll co = 0, sum = 0;
+		cliques(ed, [&](B x) { co++; sum += (ll)x.count(); });
 		cout << co << ' ' << (double)sum / co << endl;
 	}
 	return 0;
